requirements.txt
Faker==17.6.0
PyYAML==6.0
